feat: 文档删除补偿机制、KB编译开关、快速分析模式及注释中文化
后端变更:
- document: 实现删除补偿机制，记录outbox状态，依次删除usearch/BM25/SQLite，校验一致性
- ingestion: 新增kb_compilation_error和kb_analysis_engine字段，支持快速分析模式(rust)
- ingestion_queue: 新增摄取队列命令
- vector_index: 增强索引操作
- raw_source/core/knowledge_graph: 新增辅助命令
- 全部Rust注释英译中

前端变更:
- Import: 新增KB编译开关，导入结果展示分析引擎和编译错误
- Settings: 新增KB编译配置项
- OutlineNode: 大纲节点交互增强
- WikiLinkEditor: Wiki链接编辑器改进
- tauri-commands/wiki-commands: 新增KB编译相关API
- 全部TS/TSX注释英译中

+++ src-tauri/src/services/context_budget.rs
@@ -60,7 +60,8 @@ impl ContextBudget {
                 let current = final_alloc.get(&claim.slot).copied().unwrap_or(0);
                 let deficit = claim.ideal_tokens.saturating_sub(current);
                 let share = if remaining > 0 {
-                    remaining * deficit / total_ideal
+                    // 转换为 u64 进行计算，防止在大上下文窗口下乘法溢出
+                    ((remaining as u64 * deficit as u64) / total_ideal as u64) as u32
                 } else {
                     0
                 };
